fix: remove inccorrect claim of SQLAlchemy 1 support
CI tests were not actually testing on SQLAlchemy 1, so this removes the claim
that SQLAlchemy 1 is supported.

The code here in pgarrow needs support of paramstyle = "numeric_dollar" (the
only supported paramstyle of PostgreSQL itself), and that was only added in SQLAlchemy
2. While not impossible to add support for SQLAlchemy 1, suspect that is a lot
of work that realistically won't be put in any time soon.